[Flink-12164][runtime] Introduce factory pattern for HeartbeatMonitor
