+++ flink-runtime/src/main/java/org/apache/flink/runtime/heartbeat/HeartbeatManagerImpl.java
@@ -28,9 +28,6 @@ import javax.annotation.concurrent.ThreadSafe;
 
 import java.util.Collection;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Heartbeat manager implementation. The heartbeat manager maintains a map of heartbeat monitors
@@ -60,7 +57,9 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 	protected final Logger log;
 
 	/** Map containing the heartbeat monitors associated with the respective resource ID. */
-	private final ConcurrentHashMap<ResourceID, HeartbeatManagerImpl.HeartbeatMonitor<O>> heartbeatTargets;
+	private final ConcurrentHashMap<ResourceID, HeartbeatMonitor<O>> heartbeatTargets;
+
+	private final HeartbeatMonitor.Factory heartbeatMonitorFactory;
 
 	/** Running state of the heartbeat manager. */
 	protected volatile boolean stopped;
@@ -71,6 +70,23 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 			HeartbeatListener<I, O> heartbeatListener,
 			ScheduledExecutor mainThreadExecutor,
 			Logger log) {
+		this(
+			heartbeatTimeoutIntervalMs,
+			ownResourceID,
+			heartbeatListener,
+			mainThreadExecutor,
+			log,
+			new HeartbeatMonitorImpl.Factory());
+	}
+
+	public HeartbeatManagerImpl(
+			long heartbeatTimeoutIntervalMs,
+			ResourceID ownResourceID,
+			HeartbeatListener<I, O> heartbeatListener,
+			ScheduledExecutor mainThreadExecutor,
+			Logger log,
+			HeartbeatMonitor.Factory heartbeatMonitorFactory) {
+
 		Preconditions.checkArgument(heartbeatTimeoutIntervalMs > 0L, "The heartbeat timeout has to be larger than 0.");
 
 		this.heartbeatTimeoutIntervalMs = heartbeatTimeoutIntervalMs;
@@ -78,6 +94,7 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 		this.heartbeatListener = Preconditions.checkNotNull(heartbeatListener, "heartbeatListener");
 		this.mainThreadExecutor = Preconditions.checkNotNull(mainThreadExecutor);
 		this.log = Preconditions.checkNotNull(log);
+		this.heartbeatMonitorFactory = heartbeatMonitorFactory;
 		this.heartbeatTargets = new ConcurrentHashMap<>(16);
 
 		stopped = false;
@@ -109,12 +126,13 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 			if (heartbeatTargets.containsKey(resourceID)) {
 				log.debug("The target with resource ID {} is already been monitored.", resourceID);
 			} else {
-				HeartbeatManagerImpl.HeartbeatMonitor<O> heartbeatMonitor = new HeartbeatManagerImpl.HeartbeatMonitor<>(
-					resourceID,
-					heartbeatTarget,
-					mainThreadExecutor,
-					heartbeatListener,
-					heartbeatTimeoutIntervalMs);
+				HeartbeatMonitor<O> heartbeatMonitor =
+					heartbeatMonitorFactory.createHeartbeatMonitor(
+						resourceID,
+						heartbeatTarget,
+						mainThreadExecutor,
+						heartbeatListener,
+						heartbeatTimeoutIntervalMs);
 
 				heartbeatTargets.put(
 					resourceID,
@@ -133,7 +151,7 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 	@Override
 	public void unmonitorTarget(ResourceID resourceID) {
 		if (!stopped) {
-			HeartbeatManagerImpl.HeartbeatMonitor<O> heartbeatMonitor = heartbeatTargets.remove(resourceID);
+			HeartbeatMonitor<O> heartbeatMonitor = heartbeatTargets.remove(resourceID);
 
 			if (heartbeatMonitor != null) {
 				heartbeatMonitor.cancel();
@@ -145,7 +163,7 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 	public void stop() {
 		stopped = true;
 
-		for (HeartbeatManagerImpl.HeartbeatMonitor<O> heartbeatMonitor : heartbeatTargets.values()) {
+		for (HeartbeatMonitor<O> heartbeatMonitor : heartbeatTargets.values()) {
 			heartbeatMonitor.cancel();
 		}
 
@@ -202,127 +220,12 @@ public class HeartbeatManagerImpl<I, O> implements HeartbeatManager<I, O> {
 
 	HeartbeatTarget<O> reportHeartbeat(ResourceID resourceID) {
 		if (heartbeatTargets.containsKey(resourceID)) {
-			HeartbeatManagerImpl.HeartbeatMonitor<O> heartbeatMonitor = heartbeatTargets.get(resourceID);
+			HeartbeatMonitor<O> heartbeatMonitor = heartbeatTargets.get(resourceID);
 			heartbeatMonitor.reportHeartbeat();
 
 			return heartbeatMonitor.getHeartbeatTarget();
 		} else {
 			return null;
-		}
-	}
-
-	//----------------------------------------------------------------------------------------------
-	// Utility classes
-	//----------------------------------------------------------------------------------------------
-
-	/**
-	 * Heartbeat monitor which manages the heartbeat state of the associated heartbeat target. The
-	 * monitor notifies the {@link HeartbeatListener} whenever it has not seen a heartbeat signal
-	 * in the specified heartbeat timeout interval. Each heartbeat signal resets this timer.
-	 *
-	 * @param <O> Type of the payload being sent to the associated heartbeat target
-	 */
-	static class HeartbeatMonitor<O> implements Runnable {
-
-		/** Resource ID of the monitored heartbeat target. */
-		private final ResourceID resourceID;
-
-		/** Associated heartbeat target. */
-		private final HeartbeatTarget<O> heartbeatTarget;
-
-		private final ScheduledExecutor scheduledExecutor;
-
-		/** Listener which is notified about heartbeat timeouts. */
-		private final HeartbeatListener<?, ?> heartbeatListener;
-
-		/** Maximum heartbeat timeout interval. */
-		private final long heartbeatTimeoutIntervalMs;
-
-		private volatile ScheduledFuture<?> futureTimeout;
-
-		private final AtomicReference<State> state = new AtomicReference<>(State.RUNNING);
-
-		private volatile long lastHeartbeat;
-
-		HeartbeatMonitor(
-			ResourceID resourceID,
-			HeartbeatTarget<O> heartbeatTarget,
-			ScheduledExecutor scheduledExecutor,
-			HeartbeatListener<?, O> heartbeatListener,
-			long heartbeatTimeoutIntervalMs) {
-
-			this.resourceID = Preconditions.checkNotNull(resourceID);
-			this.heartbeatTarget = Preconditions.checkNotNull(heartbeatTarget);
-			this.scheduledExecutor = Preconditions.checkNotNull(scheduledExecutor);
-			this.heartbeatListener = Preconditions.checkNotNull(heartbeatListener);
-
-			Preconditions.checkArgument(heartbeatTimeoutIntervalMs > 0L, "The heartbeat timeout interval has to be larger than 0.");
-			this.heartbeatTimeoutIntervalMs = heartbeatTimeoutIntervalMs;
-
-			lastHeartbeat = 0L;
-
-			resetHeartbeatTimeout(heartbeatTimeoutIntervalMs);
-		}
-
-		HeartbeatTarget<O> getHeartbeatTarget() {
-			return heartbeatTarget;
-		}
-
-		ResourceID getHeartbeatTargetId() {
-			return resourceID;
-		}
-
-		public long getLastHeartbeat() {
-			return lastHeartbeat;
-		}
-
-		void reportHeartbeat() {
-			lastHeartbeat = System.currentTimeMillis();
-			resetHeartbeatTimeout(heartbeatTimeoutIntervalMs);
-		}
-
-		void resetHeartbeatTimeout(long heartbeatTimeout) {
-			if (state.get() == State.RUNNING) {
-				cancelTimeout();
-
-				futureTimeout = scheduledExecutor.schedule(this, heartbeatTimeout, TimeUnit.MILLISECONDS);
-
-				// Double check for concurrent accesses (e.g. a firing of the scheduled future)
-				if (state.get() != State.RUNNING) {
-					cancelTimeout();
-				}
-			}
-		}
-
-		void cancel() {
-			// we can only cancel if we are in state running
-			if (state.compareAndSet(State.RUNNING, State.CANCELED)) {
-				cancelTimeout();
-			}
-		}
-
-		private void cancelTimeout() {
-			if (futureTimeout != null) {
-				futureTimeout.cancel(true);
-			}
-		}
-
-		public boolean isCanceled() {
-			return state.get() == State.CANCELED;
-		}
-
-		@Override
-		public void run() {
-			// The heartbeat has timed out if we're in state running
-			if (state.compareAndSet(State.RUNNING, State.TIMEOUT)) {
-				heartbeatListener.notifyHeartbeatTimeout(resourceID);
-			}
-		}
-
-		private enum State {
-			RUNNING,
-			TIMEOUT,
-			CANCELED
 		}
 	}
 }
